Use tl::expected to handle and propogate errors in the shader program code
Previously, the shader program code had a lot of TODOs for error handling.
This commit implements proper error handling using [`tl::expected`](https://github.com/TartanLlama/expected),
an implementation of the `std::expected` proposal. This avoids exceptions
entirely.

To make this work, I had to create a factory function `createShaderProgram()`
that returns either a pointer to a `ShaderProgram` (now, an interface) or an
instance of `ShaderProgamError` which describes the problem. Main.cpp properly
deals with the error case by printing the error message and terminating. In
the success case, the pointer returned points to a `ShaderProgramImpl` instance.
For now, this just wraps the program id and has a `use()` method but it will
later have more methods to set uniforms etc.

Tested by messing with the shader code and filenames and saw the appropriate
errors.