Create a generic Error class
Currently, the only thing in the library is the ShaderProgram class
and when I wrote it, I created a ShaderProgramError to go with it.
I'm about to write a Texture library and will need errors for that.
I don't think I need a type system for errors (yet) so rather than
defining a TextureLoadError, I'm just going to use one Error class
for everything. This first commit creates that generic Error class
and makes ShaderProgram use it.